removed create server modal and moved styling
