+++ react-app/src/components/Servers/Servers.module.css
@@ -12,7 +12,7 @@
     justify-content: center;
 }
 
-.add {
+.button {
     display: flex;
     justify-content: center;
     align-items: center;
